Fix CEP Issue 1096
See the following link for the related Jira issue:

https://wso2.org/jira/browse/CEP-1096

The CEP server failed to shutdown because `getDataEndpoint` function continue to
busy wait for active data end points inside `onEvent` method of the
`EventQueueWorker` (which implements `EventHadler`) class even when a graceful
shutdown is being called.

To let `getDataEndpoint` method know that a shutdown is in progress, during a
shutdown, we first shutdown the reconnection service (before shutting down the
eventQueue which can't complete because of the busy waiting) and check inside
the `getDataEndpoint` method the state of reconnection service and return if it
is shutdown.