Replace elementArray with elementUtils
Instead of creating a separate object for an array, now use the default
array in javascript to store data. Inside the elementUtils now has the
all the methods which were included in the previous elementArray
object prototype. According to previous implementation array list was
shown as {0: {}, 1: {},...} . Now ot it shown as [{}, {},...] .