Change the element id generation method
Id generation for each element was redefined in here. Earlier element id
was generated like 1, 2,..., etc. With this approach each tab has
element ids starting from 1. When jsPlumb tries to render the graph from
the JSON sent by the backend, it gets messed up(this happens when two or
more tabs are open and all/some of them opened in the design view)
because it cannot uniquely identify element on the dom.(Element with the
id 1 is available in the tab one and tab 2 jsPlumb cannot identify the
correct one) In order to resolve this a unique id was given for all the
elements in the dom.( irrespective of the tab). So now element ids will
be generated as follows.

Ex:
Tab 1 --> view1_element_1,view1_element_2, ...

Tab 2 --> view2_element_1,view2_element_2, ...