Send a copy of the current configuration data to backend
By sending a copy of the current configuration data, it will not change
the current configuration.(because when sending the data to the backend
unnecessary fields for backend wll be cleaned. Then when the backend
throws an error still the user deal with the design-view and the fields
required for the front end will be lost. In order to prevent this we
send a copy of data)