Change the template for query projection
This commit is done to have separate templates for aggregation and query projections. For aggregation any number of user defined attributes can be specified and the as value is of the user choice therefore a normal input field with autocompletion is rendered for the user to type in the attribute expression and as. Along with it a + attribute button is added. For query projection the as value is pre populated with the connected element's attributes which are uneditable.