Fix for join-queries
If the join's left and right are connected to the same source, the first and second connected source doesn't get updated accordingly. I added an if condition to check if the left and right are from the same connected source, if so set the first and second connected source with the same connected source.